pyproject.toml
[project]
name = "postgres"
version = "0.1.0"
description = "I am learning Postgres with Python"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
]
